Moved nested types
