+++ App/HejDTO/Sources/HejDTO/Contests/GET/ContestDetail/Contests.Details.swift
@@ -16,9 +16,9 @@ extension Contests {
         public struct Award: Codable, Equatable {
             public let title, slug, content: String
             public let position: Int
-            public let images: [AwardImage]
+            public let images: [Image]
             
-            public struct AwardImage: Codable, Equatable {
+            public struct Image: Codable, Equatable {
                 public let alt, uuid: String
                 public let position: Int
             }
