+++ App/HejDTO/Tests/HejDTOTests/Contests/GET/ContestsDetails.Tests.swift
@@ -94,7 +94,7 @@ final class ContestsDetailsTests: XCTestCase {
                 slug: .string,
                 content: .string,
                 position: 0,
-                images: [Contests.Detail.Award.AwardImage(
+                images: [Contests.Detail.Award.Image(
                     alt: .string,
                     uuid: .string,
                     position: 0)])],
